pyproject.toml
[project]
name = "attention-for-dave2"
version = "0.1.0"
description = ""
authors = [
    {name = "Sinsankio",email = "sinethsankalpabkss@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "datasets (>=4.0.0,<5.0.0)",
    "transformers (>=4.56.0,<5.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "accelerate (>=1.10.1,<2.0.0)",
    "torchviz (>=0.0.3,<0.0.4)",
    "moviepy (>=2.2.1,<3.0.0)",
]
package-mode = false

[tool.poetry]
packages = [{include = "attention_for_dave2", from = "src"}]


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
lab = "^8.4"

[tool.poetry.group.app.dependencies]
python-socketio = "4.6.1"
python-engineio = "3.13.2"
eventlet = "^0"
flask = "^2.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
